requirements.txt
boto3==1.24.4
botocore==1.27.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
decorator==4.4.2
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
jmespath==1.0.0
moviepy==2.0.0.dev2
numpy==1.22.4
pandas==1.4.2
Pillow==9.1.1
proglog==0.1.10
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
s3transfer==0.6.0
six==1.16.0
tqdm==4.64.0
urllib3==1.26.9
